Wrap password options in a struct
