+++ iostest/src/lib.rs
@@ -62,7 +62,7 @@ fn test_round_trip_empty_generic_password() {
     println!("test_round_trip_empty_generic_password: start");
     let name = "test_empty_generic_password_input";
     let in_pass = "".as_bytes();
-    set_generic_password(name, name, in_pass, None).unwrap();
+    set_generic_password(name, name, in_pass).unwrap();
     let out_pass = get_generic_password(name, name).unwrap();
     assert_eq!(in_pass, out_pass);
     delete_generic_password(name, name).unwrap();
@@ -73,7 +73,7 @@ fn test_round_trip_ascii_generic_password() {
     println!("test_round_trip_ascii_generic_password: start");
     let name = "test_round_trip_ascii_generic_password";
     let password = "test ascii password".as_bytes();
-    set_generic_password(name, name, password, None).unwrap();
+    set_generic_password(name, name, password).unwrap();
     let stored_password = get_generic_password(name, name).unwrap();
     assert_eq!(stored_password, password);
     delete_generic_password(name, name).unwrap();
@@ -84,7 +84,7 @@ fn test_round_trip_non_ascii_generic_password() {
     println!("test_round_trip_non_ascii_generic_password: start");
     let name = "test_round_trip_non_ascii_generic_password";
     let password = "このきれいな花は桜です".as_bytes();
-    set_generic_password(name, name, password, None).unwrap();
+    set_generic_password(name, name, password).unwrap();
     let stored_password = get_generic_password(name, name).unwrap();
     assert_eq!(stored_password, password);
     delete_generic_password(name, name).unwrap();
@@ -95,7 +95,7 @@ fn test_round_trip_non_utf8_generic_password() {
     println!("test_round_trip_non_utf8_generic_password: start");
     let name = "test_round_trip_non_utf8_generic_password";
     let password: [u8; 10] = [0, 121, 122, 123, 40, 50, 126, 127, 8, 9];
-    set_generic_password(name, name, &password, None).unwrap();
+    set_generic_password(name, name, &password).unwrap();
     let stored_password = get_generic_password(name, name).unwrap();
     assert_eq!(stored_password, password);
     delete_generic_password(name, name).unwrap();
@@ -106,11 +106,11 @@ fn test_update_generic_password() {
     println!("test_update_generic_password: start");
     let name = "test_update_generic_password";
     let password = "test ascii password".as_bytes();
-    set_generic_password(name, name, password, None).unwrap();
+    set_generic_password(name, name, password).unwrap();
     let stored_password = get_generic_password(name, name).unwrap();
     assert_eq!(stored_password, password);
     let password = "このきれいな花は桜です".as_bytes();
-    set_generic_password(name, name, password, None).unwrap();
+    set_generic_password(name, name, password).unwrap();
     let stored_password = get_generic_password(name, name).unwrap();
     assert_eq!(stored_password, password);
     delete_generic_password(name, name).unwrap();
@@ -145,7 +145,7 @@ fn test_round_trip_empty_internet_password() {
     println!("test_round_trip_empty_internet_password: start");
     let name = "test_empty_internet_password_input";
     let in_pass = "".as_bytes();
-    set_internet_password(name, None, name, "/test", None, HTTP, Any, in_pass, None).unwrap();
+    set_internet_password(name, None, name, "/test", None, HTTP, Any, in_pass).unwrap();
     let out_pass = get_internet_password(name, None, name, "/test", None, HTTP, Any).unwrap();
     assert_eq!(in_pass, out_pass);
     delete_internet_password(name, None, name, "/test", None, HTTP, Any).unwrap();
@@ -156,7 +156,7 @@ fn test_round_trip_ascii_internet_password() {
     println!("test_round_trip_ascii_internet_password: start");
     let name = "test_round_trip_ascii_internet_password";
     let password = "test ascii password".as_bytes();
-    set_internet_password(name, None, name, "/test", None, HTTP, Any, password, None).unwrap();
+    set_internet_password(name, None, name, "/test", None, HTTP, Any, password).unwrap();
     let stored_password = get_internet_password(name, None, name, "/test", None, HTTP, Any).unwrap();
     assert_eq!(stored_password, password);
     delete_internet_password(name, None, name, "/test", None, HTTP, Any).unwrap();
@@ -167,7 +167,7 @@ fn test_round_trip_non_ascii_internet_password() {
     println!("test_round_trip_non_ascii_internet_password: start");
     let name = "test_round_trip_non_ascii_internet_password";
     let password = "このきれいな花は桜です".as_bytes();
-    set_internet_password(name, None, name, "/test", None, HTTP, Any, password, None).unwrap();
+    set_internet_password(name, None, name, "/test", None, HTTP, Any, password).unwrap();
     let stored_password = get_internet_password(name, None, name, "/test", None, HTTP, Any).unwrap();
     assert_eq!(stored_password, password);
     delete_internet_password(name, None, name, "/test", None, HTTP, Any).unwrap();
@@ -178,7 +178,7 @@ fn test_round_trip_non_utf8_internet_password() {
     println!("test_round_trip_non_utf8_internet_password: start");
     let name = "test_round_trip_non_utf8_internet_password";
     let password: [u8; 10] = [0, 121, 122, 123, 40, 50, 126, 127, 8, 9];
-    set_internet_password(name, None, name, "/test", None, HTTP, Any, &password, None).unwrap();
+    set_internet_password(name, None, name, "/test", None, HTTP, Any, &password).unwrap();
     let stored_password = get_internet_password(name, None, name, "/test", None, HTTP, Any).unwrap();
     assert_eq!(stored_password, password);
     delete_internet_password(name, None, name, "/test", None, HTTP, Any).unwrap();
@@ -189,11 +189,11 @@ fn test_update_internet_password() {
     println!("test_update_internet_password: start");
     let name = "test_update_internet_password";
     let password = "test ascii password".as_bytes();
-    set_internet_password(name, None, name, "/test", None, HTTP, Any, password, None).unwrap();
+    set_internet_password(name, None, name, "/test", None, HTTP, Any, password).unwrap();
     let stored_password = get_internet_password(name, None, name, "/test", None, HTTP, Any).unwrap();
     assert_eq!(stored_password, password);
     let password = "このきれいな花は桜です".as_bytes();
-    set_internet_password(name, None, name, "/test", None, HTTP, Any, password, None).unwrap();
+    set_internet_password(name, None, name, "/test", None, HTTP, Any, password).unwrap();
     let stored_password = get_internet_password(name, None, name, "/test", None, HTTP, Any).unwrap();
     assert_eq!(stored_password, password);
     delete_internet_password(name, None, name, "/test", None, HTTP, Any).unwrap();
