+++ security-framework/src/passwords.rs
@@ -4,23 +4,17 @@
 //! version of these functions in the macOS extensions module.
 
 use crate::base::Result;
+use crate::passwords_options::PasswordOptions;
 use crate::{cvt, Error};
-use crate::access_control::SecAccessControl;
-use core_foundation::base::{CFType, TCFType, CFOptionFlags};
+use core_foundation::base::{TCFType};
 use core_foundation::boolean::CFBoolean;
 use core_foundation::data::CFData;
 use core_foundation::dictionary::CFDictionary;
-use core_foundation::number::CFNumber;
 use core_foundation::string::CFString;
 use core_foundation_sys::base::{CFGetTypeID, CFRelease, CFTypeRef};
 use core_foundation_sys::data::CFDataRef;
-use security_framework_sys::access_control::*;
 use security_framework_sys::base::{errSecDuplicateItem, errSecParam};
-use security_framework_sys::item::{
-    kSecAttrAccessControl, kSecAttrAccount, kSecAttrAuthenticationType, kSecAttrPath, kSecAttrPort, kSecAttrProtocol,
-    kSecAttrSecurityDomain, kSecAttrServer, kSecAttrService, kSecClass, kSecClassGenericPassword,
-    kSecClassInternetPassword, kSecReturnData, kSecValueData,
-};
+use security_framework_sys::item::{kSecReturnData, kSecValueData,};
 use security_framework_sys::keychain::{SecAuthenticationType, SecProtocolType};
 use security_framework_sys::keychain_item::{
     SecItemAdd, SecItemCopyMatching, SecItemDelete, SecItemUpdate,
@@ -28,20 +22,20 @@ use security_framework_sys::keychain_item::{
 
 /// Set a generic password for the given service and account.
 /// Creates or updates a keychain entry.
-pub fn set_generic_password(service: &str, account: &str, password: &[u8], access_control_options: Option<AccessControlOptions>) -> Result<()> {
-    let mut query = generic_password_query(service, account);
-    set_password_internal(&mut query, password, access_control_options)
+pub fn set_generic_password(service: &str, account: &str, password: &[u8]) -> Result<()> {
+    let mut options = PasswordOptions::new_generic_password(service, account);
+    set_password_internal(&mut options, password)
 }
 
 /// Get the generic password for the given service and account.  If no matching
 /// keychain entry exists, fails with error code `errSecItemNotFound`.
 pub fn get_generic_password(service: &str, account: &str) -> Result<Vec<u8>> {
-    let mut query = generic_password_query(service, account);
-    query.push((
+    let mut options = PasswordOptions::new_generic_password(service, account);
+    options.query.push((
         unsafe { CFString::wrap_under_get_rule(kSecReturnData) },
         CFBoolean::from(true).as_CFType(),
     ));
-    let params = CFDictionary::from_CFType_pairs(&query);
+    let params = CFDictionary::from_CFType_pairs(&options.query);
     let mut ret: CFTypeRef = std::ptr::null();
     cvt(unsafe { SecItemCopyMatching(params.as_concrete_TypeRef(), &mut ret) })?;
     get_password_and_release(ret)
@@ -50,8 +44,8 @@ pub fn get_generic_password(service: &str, account: &str) -> Result<Vec<u8>> {
 /// Delete the generic password keychain entry for the given service and account.
 /// If none exists, fails with error code `errSecItemNotFound`.
 pub fn delete_generic_password(service: &str, account: &str) -> Result<()> {
-    let query = generic_password_query(service, account);
-    let params = CFDictionary::from_CFType_pairs(&query);
+    let options = PasswordOptions::new_generic_password(service, account);
+    let params = CFDictionary::from_CFType_pairs(&options.query);
     cvt(unsafe { SecItemDelete(params.as_concrete_TypeRef()) })
 }
 
@@ -66,19 +60,18 @@ pub fn set_internet_password(
     port: Option<u16>,
     protocol: SecProtocolType,
     authentication_type: SecAuthenticationType,
-    password: &[u8],
-    access_control_options: Option<AccessControlOptions>,
+    password: &[u8]
 ) -> Result<()> {
-    let mut query = internet_password_query(
+    let mut options = PasswordOptions::new_internet_password(
         server,
         security_domain,
         account,
         path,
         port,
         protocol,
-        authentication_type,
+        authentication_type
     );
-    set_password_internal(&mut query, password, access_control_options)
+    set_password_internal(&mut options, password)
 }
 
 /// Get the internet password for the given endpoint parameters.  If no matching
@@ -92,7 +85,7 @@ pub fn get_internet_password(
     protocol: SecProtocolType,
     authentication_type: SecAuthenticationType,
 ) -> Result<Vec<u8>> {
-    let mut query = internet_password_query(
+    let mut options = PasswordOptions::new_internet_password(
         server,
         security_domain,
         account,
@@ -101,11 +94,11 @@ pub fn get_internet_password(
         protocol,
         authentication_type,
     );
-    query.push((
+    options.query.push((
         unsafe { CFString::wrap_under_get_rule(kSecReturnData) },
         CFBoolean::from(true).as_CFType(),
     ));
-    let params = CFDictionary::from_CFType_pairs(&query);
+    let params = CFDictionary::from_CFType_pairs(&options.query);
     let mut ret: CFTypeRef = std::ptr::null();
     cvt(unsafe { SecItemCopyMatching(params.as_concrete_TypeRef(), &mut ret) })?;
     get_password_and_release(ret)
@@ -122,7 +115,7 @@ pub fn delete_internet_password(
     protocol: SecProtocolType,
     authentication_type: SecAuthenticationType,
 ) -> Result<()> {
-    let query = internet_password_query(
+    let options = PasswordOptions::new_internet_password(
         server,
         security_domain,
         account,
@@ -131,128 +124,25 @@ pub fn delete_internet_password(
         protocol,
         authentication_type,
     );
-    let params = CFDictionary::from_CFType_pairs(&query);
+    let params = CFDictionary::from_CFType_pairs(&options.query);
     cvt(unsafe { SecItemDelete(params.as_concrete_TypeRef()) })
-}
-
-// Generic passwords are identified by service and account.  They have other
-// attributes, but this interface doesn't allow specifying them.
-fn generic_password_query(service: &str, account: &str) -> Vec<(CFString, CFType)> {
-    let query = vec![
-        (
-            unsafe { CFString::wrap_under_get_rule(kSecClass) },
-            unsafe { CFString::wrap_under_get_rule(kSecClassGenericPassword).as_CFType() },
-        ),
-        (
-            unsafe { CFString::wrap_under_get_rule(kSecAttrService) },
-            CFString::from(service).as_CFType(),
-        ),
-        (
-            unsafe { CFString::wrap_under_get_rule(kSecAttrAccount) },
-            CFString::from(account).as_CFType(),
-        ),
-    ];
-    query
-}
-
-// Internet passwords are identified by a number of attributes.
-// They can have others, but this interface doesn't allow specifying them.
-fn internet_password_query(
-    server: &str,
-    security_domain: Option<&str>,
-    account: &str,
-    path: &str,
-    port: Option<u16>,
-    protocol: SecProtocolType,
-    authentication_type: SecAuthenticationType,
-) -> Vec<(CFString, CFType)> {
-    let mut query = vec![
-        (
-            unsafe { CFString::wrap_under_get_rule(kSecClass) },
-            unsafe { CFString::wrap_under_get_rule(kSecClassInternetPassword) }.as_CFType(),
-        ),
-        (
-            unsafe { CFString::wrap_under_get_rule(kSecAttrServer) },
-            CFString::from(server).as_CFType(),
-        ),
-        (
-            unsafe { CFString::wrap_under_get_rule(kSecAttrPath) },
-            CFString::from(path).as_CFType(),
-        ),
-        (
-            unsafe { CFString::wrap_under_get_rule(kSecAttrAccount) },
-            CFString::from(account).as_CFType(),
-        ),
-        (
-            unsafe { CFString::wrap_under_get_rule(kSecAttrProtocol) },
-            CFNumber::from(protocol as i32).as_CFType(),
-        ),
-        (
-            unsafe { CFString::wrap_under_get_rule(kSecAttrAuthenticationType) },
-            CFNumber::from(authentication_type as i32).as_CFType(),
-        ),
-    ];
-    if let Some(domain) = security_domain {
-        query.push((
-            unsafe { CFString::wrap_under_get_rule(kSecAttrSecurityDomain) },
-            CFString::from(domain).as_CFType(),
-        ))
-    }
-    if let Some(port) = port {
-        query.push((
-            unsafe { CFString::wrap_under_get_rule(kSecAttrPort) },
-            CFNumber::from(i32::from(port)).as_CFType(),
-        ))
-    }
-    query
-}
-
-bitflags::bitflags! {
-    /// The option flags used to configure the evaluation of a `SecAccessControl`.
-    pub struct AccessControlOptions: CFOptionFlags {
-        /** Constraint to access an item with either biometry or passcode. */
-        const USER_PRESENCE = kSecAccessControlUserPresence;
-        /** Constraint to access an item with Touch ID for any enrolled fingers, or Face ID. */
-        const BIOMETRY_ANY = kSecAccessControlBiometryAny;
-        /** Constraint to access an item with Touch ID for currently enrolled fingers, or from Face ID with the currently enrolled user. */
-        const BIOMETRY_CURRENT_SET = kSecAccessControlBiometryCurrentSet;
-        /** Constraint to access an item with a passcode. */
-        const DEVICE_PASSCODE = kSecAccessControlDevicePasscode;
-        /** Constraint to access an item with a watch. */
-        const WATCH = kSecAccessControlWatch;
-        /** Indicates that at least one constraint must be satisfied. */
-        const OR = kSecAccessControlOr;
-        /** Indicates that all constraints must be satisfied. */
-        const AND = kSecAccessControlAnd;
-        /** Enable a private key to be used in signing a block of data or verifying a signed block. */
-        const PRIVATE_KEY_USAGE = kSecAccessControlPrivateKeyUsage;
-        /** Option to use an application-provided password for data encryption key generation. */
-        const APPLICATION_PASSWORD = kSecAccessControlApplicationPassword;
-    }
 }
 
 // This starts by trying to create the password with the given query params.
 // If the creation attempt reveals that one exists, its password is updated.
-fn set_password_internal(query: &mut Vec<(CFString, CFType)>, password: &[u8], access_control_options: Option<AccessControlOptions>) -> Result<()> {
-    let query_len = query.len();
-    query.push((
+fn set_password_internal(options: &mut PasswordOptions, password: &[u8]) -> Result<()> {
+    let query_len = options.query.len();
+    options.query.push((
         unsafe { CFString::wrap_under_get_rule(kSecValueData) },
         CFData::from_buffer(password).as_CFType(),
     ));
-
-    if let Some(options) = access_control_options {
-        query.push((
-            unsafe { CFString::wrap_under_get_rule(kSecAttrAccessControl) },
-            SecAccessControl::create_with_flags(options.bits()).unwrap().as_CFType()
-        ));
-    }
     
-    let params = CFDictionary::from_CFType_pairs(query);
+    let params = CFDictionary::from_CFType_pairs(&options.query);
     let mut ret = std::ptr::null();
     let status = unsafe { SecItemAdd(params.as_concrete_TypeRef(), &mut ret) };
     if status == errSecDuplicateItem {
-        let params = CFDictionary::from_CFType_pairs(&query[0..query_len]);
-        let update = CFDictionary::from_CFType_pairs(&query[query_len..]);
+        let params = CFDictionary::from_CFType_pairs(&options.query[0..query_len]);
+        let update = CFDictionary::from_CFType_pairs(&options.query[query_len..]);
         cvt(unsafe { SecItemUpdate(params.as_concrete_TypeRef(), update.as_concrete_TypeRef()) })
     } else {
         cvt(status)
@@ -313,7 +203,7 @@ mod test {
     #[test]
     fn roundtrip_generic() {
         let name = "roundtrip_generic";
-        set_generic_password(name, name, name.as_bytes(), None).expect("set_generic_password");
+        set_generic_password(name, name, name.as_bytes()).expect("set_generic_password");
         let pass = get_generic_password(name, name).expect("get_generic_password");
         assert_eq!(name.as_bytes(), pass);
         delete_generic_password(name, name).expect("delete_generic_password")
@@ -322,9 +212,9 @@ mod test {
     #[test]
     fn update_generic() {
         let name = "update_generic";
-        set_generic_password(name, name, name.as_bytes(), None).expect("set_generic_password");
+        set_generic_password(name, name, name.as_bytes()).expect("set_generic_password");
         let alternate = "update_generic_alternate";
-        set_generic_password(name, name, alternate.as_bytes(), None).expect("set_generic_password");
+        set_generic_password(name, name, alternate.as_bytes()).expect("set_generic_password");
         let pass = get_generic_password(name, name).expect("get_generic_password");
         assert_eq!(pass, alternate.as_bytes());
         delete_generic_password(name, name).expect("delete_generic_password")
@@ -389,15 +279,14 @@ mod test {
     #[test]
     fn roundtrip_internet() {
         let name = "roundtrip_internet";
-        let (server, domain, account, path, port, protocol, auth, access) = (
+        let (server, domain, account, path, port, protocol, auth) = (
             name,
             None,
             name,
             "/",
             Some(8080u16),
             SecProtocolType::HTTP,
-            SecAuthenticationType::Any,
-            None,
+            SecAuthenticationType::Any
         );
         set_internet_password(
             server,
@@ -407,8 +296,7 @@ mod test {
             port,
             protocol,
             auth,
-            name.as_bytes(),
-            access,
+            name.as_bytes()
         )
         .expect("set_internet_password");
         let pass = get_internet_password(server, domain, account, path, port, protocol, auth)
@@ -421,15 +309,14 @@ mod test {
     #[test]
     fn update_internet() {
         let name = "update_internet";
-        let (server, domain, account, path, port, protocol, auth, access) = (
+        let (server, domain, account, path, port, protocol, auth) = (
             name,
             None,
             name,
             "/",
             Some(8080u16),
             SecProtocolType::HTTP,
-            SecAuthenticationType::Any,
-            None,
+            SecAuthenticationType::Any
         );
         set_internet_password(
             server,
@@ -439,8 +326,7 @@ mod test {
             port,
             protocol,
             auth,
-            name.as_bytes(),
-            access
+            name.as_bytes()
         )
         .expect("set_internet_password");
         let alternate = "alternate_internet_password";
@@ -452,8 +338,7 @@ mod test {
             port,
             protocol,
             auth,
-            alternate.as_bytes(),
-            access
+            alternate.as_bytes()
         )
         .expect("set_internet_password");
         let pass = get_internet_password(server, domain, account, path, port, protocol, auth)
